unified-api common

+++ mods/service/mqttd/mqtt_test.go
@@ -408,7 +408,7 @@ func (tm *TestMock) NewDB() *dbMock {
 				return &RowMock{
 					ErrFunc: func() error { return nil },
 					ScanFunc: func(cols ...any) error {
-						*(cols[0].(*int)) = 0                          // TABLE_ID
+						*(cols[0].(*int64)) = 0                        // TABLE_ID
 						*(cols[1].(*api.TableType)) = api.TableTypeTag // TABLE_TYPE
 						*(cols[3].(*int)) = 3                          // TABLE_COLCOUNT
 						return nil
